# [PERF] account: speed up _inverse_product_id