Currently the condition on `self.product_id or not self.account_id` is inside the lambda passed to `_conditional_add_to_compute`. `_conditional_add_to_compute` calls the lambda in a filtered so this condition will be evaluated for all the lines in self. As it only depends on self, i.e. not the current line, we can move it outside of the implicit for-loop.

#### speedup

Customer database. Calling `_inverse_product_id` increasing the cardinality of self.

| self size | Before PR | After PR |
|:----------:|:---------------:|:-----------:|
|     10    |    5ms    |   5ms    |
|     100   |    14ms   |   11ms   |
|     1000  |    271ms  |   36ms   |
|     10000 |    28.7s  |   300ms  |
|     50000 |    16min  |   1.5s   |

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr