@robodoo 2024-06-26
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/170934.png)](https://mergebot.odoo.com/odoo/odoo/pull/170934)